refactor: decompose ResumePreviewPage and SkillGroups into folders
큰 파일 두 개를 각각 폴더로 분해해 책임을 분리하고 가독성을 높임. 동작 변경 없음.

### ResumePreviewPage (1219 → 1075 메인 + 4 헬퍼 파일)
- src/pages/ResumePreviewPage.tsx → src/pages/ResumePreviewPage/index.tsx (이동, git mv)
- helpers.ts: linkOrder/projectLinkOrder, cloneResumeData, groupBlocksBySection
- editor.tsx: SectionHeading, FieldLabel, CollapsibleEditorSection, CollapsibleSubItem, InputField, TextAreaField
- preview.tsx: DetailPreviewItem
- useSidebarPin.ts: 사이드바 고정 상태 + ref + localStorage sync

### SkillGroups (434 → 23 메인 + 6 컴포넌트)
- src/components/AiDevKitModal/SkillGroups.tsx → SkillGroups/index.tsx (이동)
- SkillFlowSteps.tsx / SkillDetailGrid.tsx / SkillModuleDiagram.tsx /
  SkillSectionCard.tsx / SkillItemCard.tsx
- icons.tsx: SectionIcon (배열 인덱싱) + ItemIcon (키워드 매칭 룩업 테이블)
  - 기존 80+ 줄 if-chain → matchItemIcon() 8줄 + 데이터 테이블로 데이터-드리븐화
  - 신규 아이콘 추가 시 keywords/Icon 한 행 추가만 필요

### 테스트
- src/components/AiDevKitModal/SkillGroups/icons.test.tsx 추가 (matchItemIcon 5개 테스트)
- 전체 290개 테스트 통과 (기존 287 + 신규 3 등)
- 프로덕션 빌드 정상
